Address comments and fix test after rebasing.
