+++ cypress.config.ts
@@ -12,7 +12,6 @@ module.exports = defineConfig({
   responseTimeout: 60000,
   viewportWidth: 2000,
   viewportHeight: 1320,
-  retries: 2,
   env: {
     ENGINE: {
       name: 'default',
