--- cypress/integration/constants.js
@@ -1,10 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export const INDEX_CLUSTER_NAME = 'data-logs-1';
-export const INDEX_NAME = 'data_logs_small_time_1';
-export const INDEX_PATTERN_NAME = 'data-logs-1::data_logs_small_time_1*';
-export const SEARCH_ABSOLUTE_START_DATE = 'Dec 31, 2020 @ 16:00:00.000';
-export const SEARCH_ABSOLUTE_END_DATE = 'Dec 31, 2022 @ 14:14:42.801';
