+++ cypress/integration/core-opensearch-dashboards/opensearch-dashboards/apps/query_enhancements/constants.js
@@ -3,5 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export const DATASOURCE_NAME = 'query-cluster';
-export const WORKSPACE_NAME = 'query-workspace';
+export const DATASOURCE_NAME = 'data-logs-1';
+export const INDEX_NAME = 'data_logs_small_time_1';
+export const DEFAULT_TIME_INDEX_PATTERN_NAME = 'data_logs_small_time*';
+export const DEFAULT_NO_TIME_INDEX_PATTERN_NAME = 'data_logs_small_no_time*';
+export const INDEX_PATTERN_NAME = `${DATASOURCE_NAME}::${DEFAULT_TIME_INDEX_PATTERN_NAME}`;
+export const SEARCH_ABSOLUTE_START_DATE = 'Dec 31, 2020 @ 16:00:00.000';
+export const SEARCH_ABSOLUTE_END_DATE = 'Dec 31, 2022 @ 14:14:42.801';
