+++ cypress/support/e2e.js
@@ -6,9 +6,6 @@
 import '../utils/commands.js';
 import '../utils/dashboards/data_explorer/commands.js';
 
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
-
 const scopedHistoryNavigationError = /^[^(ScopedHistory instance has fell out of navigation scope)]/;
 Cypress.on('uncaught:exception', (err) => {
   /* returning false here prevents Cypress from failing the test */
