+++ cypress/utils/dashboards/workspace-plugin/constants.js
@@ -3,4 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export const WORKSPACE_API_PREFIX = '/api/workspaces';
+export const PATHS = {
+  WORKSPACE_API_PREFIX: '/api/workspaces',
+};
